Added testing
Added testing for remove method to ensure element was removed and no null spaces were left.  Testing for get size method, to test the method returned the size of the actual number of elements stored.  add(index, item) method to check for null value at index.  If not null, adding item to that index and shifting existing elements right one index as needed